[Feat] #61 - 스플래쉬 뷰 구현
