+++ Smeem-iOS/Smeem-iOS/Supports/SceneDelegate.swift
@@ -17,7 +17,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let scene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: scene)
         
-        let rootViewController = UINavigationController(rootViewController: UserNicknameViewController())
+        let rootViewController = UINavigationController(rootViewController: SplashViewController())
         self.window?.rootViewController = rootViewController
         self.window?.makeKeyAndVisible()
     }
